Re-add cakephp/bake to non-dev dependencies (task #8927)
We usually install `cakephp/bake` as developer requirement on all
development and test environments.  In production, however, running
`composer install --no-dev` skips bake altogether.  This is fine for
most other projects, but not for us.

One of our migration scripts (`src/Shell/Task/UpgradeFileTask.php`)
requires the presence of bake and breaks the deployment during the
`app:update` if it's not there.