Adjusted assertion (task #8712)
Since we are testing beforeLookup event, we cannot rely on pagination limit, which is only applied later on. Test was passing with the old assertion because the users fixture includes exactly 10 records. But if a record is added or removed it will fail. We now simply assert that the query is not empty.